Add defaulted copy functions when implicit is deprecated
C++11 deprecated the implicit generation of copy constructors and
move assignment operators if the other, or the destructor, is user
declared.  Modern gcc compilers warn about this deprecation, so
this patch resolves the simple cases that can be handled with an
'= default' inline definition and defers to a follow-up patch any
cases that require more than a little thought.

It incorporates drive-by fixes to clean up several test drivers
that are touched by this patch, and one required #include for a
test driver to compile with the latest gcc standard library.